requirements.txt
Flask==1.1.2
#mod_wsgi
flask-mysql
